parse_userstories/steps/test_steps_bdd.py
# Auto-generated pytest-bdd step definitions.
# Run: pytest -k feature
import pytest
from pytest_bdd import given, when, then, scenarios

# Link feature(s)
scenarios("features/*.feature")

# Example shared test data (E2E placeholders)
TEST_CONTEXT = {
    "patient_id": "PAT-001",
    "session_id": "SES-001",
    "clinician_id": "DOC-123",
}

@given("I am logged into the Clinician Portal.")
def step_given():
    # TODO: implement setup for: I am logged into the Clinician Portal.
    # e.g., create patient in DB using TEST_CONTEXT["patient_id"]
    pass

@when("I search for a patient's record.")
def step_when():
    # TODO: implement action for: I search for a patient's record.
    # e.g., call API to sign in/out, upload document, etc.
    pass

@then("I can view digitized assessments and care plans for that patient.")
def step_then():
    # TODO: implement assertion for: I can view digitized assessments and care plans for that patient.
    # e.g., assert response.status_code == 200 or record exists in DB
    pass
