parse_userstories/requirements.txt
pdfminer.six
python-docx
jsonschema
pydantic
google-cloud-aiplatform>=1.71.0
vertexai>=1.71.0
tqdm